requirements.txt
# One Pro IMU Retriever Demo
numpy>=1.21.0
pygame>=2.1.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6


